requirements.txt
Flask
python-dotenv
pydantic
google-genai